+++ sentry-core/src/client.rs
@@ -1,5 +1,4 @@
 use std::borrow::Cow;
-use std::ffi::{OsStr, OsString};
 use std::fmt;
 use std::sync::Arc;
 use std::sync::RwLock;
@@ -11,12 +10,17 @@ use crate::backtrace_support::process_event_stacktrace;
 pub use crate::clientoptions::ClientOptions;
 use crate::constants::SDK_INFO;
 use crate::hub::Hub;
-use crate::internals::{Dsn, ParseDsnError, Uuid};
+use crate::internals::{Dsn, Uuid};
 use crate::protocol::{DebugMeta, Event};
 use crate::scope::Scope;
 use crate::transport::Transport;
 use crate::utils;
 
+impl<T: Into<ClientOptions>> From<T> for Client {
+    fn from(o: T) -> Client {
+        Client::with_options(o.into())
+    }
+}
 /// The Sentry client object.
 pub struct Client {
     options: ClientOptions,
@@ -38,99 +42,6 @@ impl Clone for Client {
             options: self.options.clone(),
             transport: RwLock::new(self.transport.read().unwrap().clone()),
         }
-    }
-}
-
-/// Helper trait to convert a string into an `Option<Dsn>`.
-///
-/// This converts a value into a DSN by parsing.  The empty string or
-/// null values result in no DSN being parsed.
-pub trait IntoDsn {
-    /// Converts the value into a `Result<Option<Dsn>, E>`.
-    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError>;
-}
-
-impl<T: Into<ClientOptions>> From<T> for Client {
-    fn from(o: T) -> Client {
-        Client::with_options(o.into())
-    }
-}
-
-impl<T: IntoDsn> From<(T, ClientOptions)> for ClientOptions {
-    fn from((into_dsn, mut opts): (T, ClientOptions)) -> ClientOptions {
-        opts.dsn = into_dsn.into_dsn().expect("invalid value for DSN");
-        opts
-    }
-}
-
-impl<T: IntoDsn> From<T> for ClientOptions {
-    fn from(into_dsn: T) -> ClientOptions {
-        ClientOptions {
-            dsn: into_dsn.into_dsn().expect("invalid value for DSN"),
-            ..ClientOptions::default()
-        }
-    }
-}
-
-impl<I: IntoDsn> IntoDsn for Option<I> {
-    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
-        match self {
-            Some(into_dsn) => into_dsn.into_dsn(),
-            None => Ok(None),
-        }
-    }
-}
-
-impl IntoDsn for () {
-    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
-        Ok(None)
-    }
-}
-
-impl<'a> IntoDsn for &'a str {
-    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
-        if self.is_empty() {
-            Ok(None)
-        } else {
-            self.parse().map(Some)
-        }
-    }
-}
-
-impl<'a> IntoDsn for Cow<'a, str> {
-    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
-        let x: &str = &self;
-        x.into_dsn()
-    }
-}
-
-impl<'a> IntoDsn for &'a OsStr {
-    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
-        self.to_string_lossy().into_dsn()
-    }
-}
-
-impl IntoDsn for OsString {
-    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
-        self.as_os_str().into_dsn()
-    }
-}
-
-impl IntoDsn for String {
-    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
-        self.as_str().into_dsn()
-    }
-}
-
-impl<'a> IntoDsn for &'a Dsn {
-    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
-        Ok(Some(self.clone()))
-    }
-}
-
-impl IntoDsn for Dsn {
-    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
-        Ok(Some(self))
     }
 }
 
@@ -390,7 +301,6 @@ impl Drop for ClientInitGuard {
 /// and to then bind it to the hub except it's also possible to directly pass
 /// a client.  For more information about the formats accepted see
 /// `Client::from_config`.
-#[cfg(feature = "with_client_implementation")]
 pub fn init<C: Into<Client>>(cfg: C) -> ClientInitGuard {
     let client = Arc::new(cfg.into());
     Hub::with(|hub| hub.bind_client(Some(client.clone())));
