+++ sentry-core/src/clientoptions.rs
@@ -1,11 +1,12 @@
 use std::borrow::Cow;
+use std::ffi::{OsStr, OsString};
 use std::fmt;
 use std::panic::RefUnwindSafe;
 use std::sync::Arc;
 use std::time::Duration;
 
 use crate::constants::USER_AGENT;
-use crate::internals::Dsn;
+use crate::internals::{Dsn, ParseDsnError};
 use crate::protocol::{Breadcrumb, Event};
 use crate::transport::{DefaultTransportFactory, TransportFactory};
 use crate::utils;
@@ -177,5 +178,92 @@ impl Default for ClientOptions {
             before_send: None,
             before_breadcrumb: None,
         }
+    }
+}
+
+/// Helper trait to convert a string into an `Option<Dsn>`.
+///
+/// This converts a value into a DSN by parsing.  The empty string or
+/// null values result in no DSN being parsed.
+pub trait IntoDsn {
+    /// Converts the value into a `Result<Option<Dsn>, E>`.
+    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError>;
+}
+
+impl<T: IntoDsn> From<(T, ClientOptions)> for ClientOptions {
+    fn from((into_dsn, mut opts): (T, ClientOptions)) -> ClientOptions {
+        opts.dsn = into_dsn.into_dsn().expect("invalid value for DSN");
+        opts
+    }
+}
+
+impl<T: IntoDsn> From<T> for ClientOptions {
+    fn from(into_dsn: T) -> ClientOptions {
+        ClientOptions {
+            dsn: into_dsn.into_dsn().expect("invalid value for DSN"),
+            ..ClientOptions::default()
+        }
+    }
+}
+
+impl<I: IntoDsn> IntoDsn for Option<I> {
+    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
+        match self {
+            Some(into_dsn) => into_dsn.into_dsn(),
+            None => Ok(None),
+        }
+    }
+}
+
+impl IntoDsn for () {
+    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
+        Ok(None)
+    }
+}
+
+impl<'a> IntoDsn for &'a str {
+    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
+        if self.is_empty() {
+            Ok(None)
+        } else {
+            self.parse().map(Some)
+        }
+    }
+}
+
+impl<'a> IntoDsn for Cow<'a, str> {
+    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
+        let x: &str = &self;
+        x.into_dsn()
+    }
+}
+
+impl<'a> IntoDsn for &'a OsStr {
+    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
+        self.to_string_lossy().into_dsn()
+    }
+}
+
+impl IntoDsn for OsString {
+    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
+        self.as_os_str().into_dsn()
+    }
+}
+
+impl IntoDsn for String {
+    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
+        self.as_str().into_dsn()
+    }
+}
+
+impl<'a> IntoDsn for &'a Dsn {
+    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
+        Ok(Some(self.clone()))
+    }
+}
+
+impl IntoDsn for Dsn {
+    fn into_dsn(self) -> Result<Option<Dsn>, ParseDsnError> {
+        Ok(Some(self))
     }
 }
